+++ src/misc/programs.cpp
@@ -465,7 +465,7 @@ void CONFIG::Run(void) {
 				WriteOut(MSG_Get("PROGRAM_CONFIG_HLP_SECTHLP"),pvars[0].c_str());
 				while(true) {
 					// list the properties
-					Property* p = psec->Get_prop(i++);
+					Property* p = psec->Get_prop((int)(i++));
 					if (p==NULL) break;
 					WriteOut("%s\n", p->propname.c_str());
 				}
@@ -473,7 +473,7 @@ void CONFIG::Run(void) {
 				// find the property by it's name
 				size_t i = 0;
 				while (true) {
-					Property *p = psec->Get_prop(i++);
+					Property *p = psec->Get_prop((int)(i++));
 					if (p==NULL) break;
 					if (!strcasecmp(p->propname.c_str(),pvars[1].c_str())) {
 						// found it; make the list of possible values
